Created separate directories to separate services and resources. Modified services files to be singletons
