+++ src/main/java/lithium/university/TwitterApplication.java
@@ -3,6 +3,7 @@ package lithium.university;
 import io.dropwizard.Application;
 import io.dropwizard.setup.Bootstrap;
 import io.dropwizard.setup.Environment;
+import lithium.university.resources.TwitterResource;
 
 
 public class TwitterApplication extends Application<TwitterConfiguration> {
